// scripts/common/web/enums/map-credit-type.enum.ts
export enum MapCreditType {
	/* Contributed significant work to the map */
	AUTHOR = 0,
	/* Did minor work on the map */
	CONTRIBUTOR = 1,
	/* Tested the map during development, or submitted reviews during map
	 * submission testing phases */
	TESTER = 2,
	/* Any miscellaneous contributions, e.g. donations, encouragement etc. */
	SPECIAL_THANKS = 3
}